Add helper class FAUtils
