+++ ocmsdk/src/main/java/com/gigigo/orchextra/core/sdk/actions/ActionHandler.java
@@ -4,7 +4,6 @@ import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
 import android.support.design.widget.Snackbar;
-import android.util.Pair;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
@@ -17,11 +16,8 @@ import com.gigigo.orchextra.core.sdk.model.detail.viewtypes.youtube.YoutubeConte
 import com.gigigo.orchextra.core.sdk.utils.DeviceUtils;
 import com.gigigo.orchextra.ocm.OCManager;
 import com.gigigo.orchextra.ocm.Ocm;
+import com.gigigo.orchextra.ocm.federatedAuth.FAUtils;
 import com.gigigo.orchextra.ocmsdk.R;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Iterator;
-import java.util.List;
 
 public class ActionHandler {
 
@@ -61,7 +57,7 @@ public class ActionHandler {
 
           Ocm.getQueryStringGenerator().createQueryString(federatedAuth, queryString -> {
             if (queryString != null && !queryString.isEmpty()) {
-              String urlWithQueryParams = addQueryParamsToUrl(queryString, url);
+              String urlWithQueryParams = FAUtils.addQueryParamsToUrl(queryString, url);
               System.out.println(
                   "Main ContentWebViewGridLayout federatedAuth url: " + urlWithQueryParams);
               if (urlWithQueryParams != null) {
@@ -91,39 +87,6 @@ public class ActionHandler {
       //todo falta que si no hay currentactivity lo lanze en webview
     }
   }
-
-  //region external browser FA
-  private static final String URL_START_QUERY_DELIMITER = "?";
-  private static final String URL_CONCAT_QUERY_DELIMITER = "&";
-  private static final String URL_QUERY_VALUE_DELIMITER = "=";
-
-  private String getQueryDelimiter(String url) {
-    try {
-      return new URL(url).getQuery() != null ? URL_CONCAT_QUERY_DELIMITER
-          : URL_START_QUERY_DELIMITER;
-    } catch (MalformedURLException e) {
-      e.printStackTrace();
-    }
-    return null;
-  }
-
-  private String addQueryParamsToUrl(List<Pair<String, String>> queryParams, String url) {
-    if (getQueryDelimiter(url) != null) {
-      url = url + getQueryDelimiter(url);
-
-      Iterator<Pair<String, String>> iterator = queryParams.iterator();
-      while (iterator.hasNext()) {
-        Pair<String, String> pair = iterator.next();
-        url =
-            url + pair.first + URL_QUERY_VALUE_DELIMITER + pair.second + URL_CONCAT_QUERY_DELIMITER;
-      }
-
-      return url.substring(0, url.length() - 2);
-    } else {
-      return null;
-    }
-  }
-  //endregion
 
   public void launchCustomTabs(String url, FederatedAuthorization federatedAuthorization) {
     if (connectionUtils.hasConnection()) {
