+++ ocmsdk/src/main/java/com/gigigo/orchextra/core/sdk/model/detail/viewtypes/WebViewContentData.java
@@ -8,7 +8,6 @@ import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.util.Log;
-import android.util.Pair;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -28,12 +27,9 @@ import com.gigigo.orchextra.core.domain.entities.elementcache.FederatedAuthoriza
 import com.gigigo.orchextra.core.sdk.ui.views.TouchyWebView;
 import com.gigigo.orchextra.ocm.OCManager;
 import com.gigigo.orchextra.ocm.Ocm;
+import com.gigigo.orchextra.ocm.federatedAuth.FAUtils;
 import com.gigigo.orchextra.ocmsdk.R;
 import java.lang.ref.WeakReference;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -42,9 +38,6 @@ public class WebViewContentData extends UiBaseContentData {
 
   private static final String EXTRA_URL = "EXTRA_URL";
   private static final String EXTRA_FEDERATED_AUTH = "EXTRA_FEDERATED_AUTH";
-  private static final String URL_START_QUERY_DELIMITER = "?";
-  private static final String URL_CONCAT_QUERY_DELIMITER = "&";
-  private static final String URL_QUERY_VALUE_DELIMITER = "=";
 
   private View mView;
   private TouchyWebView webView;
@@ -204,33 +197,6 @@ public class WebViewContentData extends UiBaseContentData {
     }
   }
 
-  private String getQueryDelimiter(String url) {
-    try {
-      return new URL(url).getQuery() != null ? URL_CONCAT_QUERY_DELIMITER
-          : URL_START_QUERY_DELIMITER;
-    } catch (MalformedURLException e) {
-      e.printStackTrace();
-    }
-    return null;
-  }
-
-  private String addQueryParamsToUrl(List<Pair<String, String>> queryParams, String url) {
-    if (getQueryDelimiter(url) != null) {
-      url = url + getQueryDelimiter(url);
-
-      Iterator<Pair<String, String>> iterator = queryParams.iterator();
-      while (iterator.hasNext()) {
-        Pair<String, String> pair = iterator.next();
-        url =
-            url + pair.first + URL_QUERY_VALUE_DELIMITER + pair.second + URL_CONCAT_QUERY_DELIMITER;
-      }
-
-      return url.substring(0, url.length() - 2);
-    } else {
-      return null;
-    }
-  }
-
   private void loadUrl() {
     showProgressView(true);
 
@@ -244,7 +210,7 @@ public class WebViewContentData extends UiBaseContentData {
           && Ocm.getQueryStringGenerator() != null) {
         Ocm.getQueryStringGenerator().createQueryString(federatedAuthorization, queryString -> {
           if (queryString != null && !queryString.isEmpty()) {
-            String urlWithQueryParams = addQueryParamsToUrl(queryString, url);
+            String urlWithQueryParams = FAUtils.addQueryParamsToUrl(queryString, url);
             //no es necesario  OCManager.saveFedexAuth(url);
             Log.d(WebViewContentData.class.getSimpleName(),
                 "federatedAuth url: " + urlWithQueryParams);
