+++ ocmsdk/src/main/java/com/gigigo/orchextra/core/sdk/utils/DeviceUtils.java
@@ -8,18 +8,14 @@ import android.net.Uri;
 import android.os.Build;
 import android.support.customtabs.CustomTabsIntent;
 import android.util.DisplayMetrics;
-import android.util.Pair;
 import android.view.Display;
 import android.view.WindowManager;
 import com.gigigo.ggglib.device.AndroidSdkVersion;
 import com.gigigo.orchextra.core.domain.entities.elementcache.FederatedAuthorization;
 import com.gigigo.orchextra.core.sdk.ui.OcmWebViewActivity;
 import com.gigigo.orchextra.ocm.Ocm;
+import com.gigigo.orchextra.ocm.federatedAuth.FAUtils;
 import com.gigigo.orchextra.ocmsdk.R;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Iterator;
-import java.util.List;
 
 public class DeviceUtils {
 
@@ -54,37 +50,6 @@ public class DeviceUtils {
       Point size = new Point();
       display.getSize(size);
       return size.x;
-    }
-  }
-
-  private static final String URL_START_QUERY_DELIMITER = "?";
-  private static final String URL_CONCAT_QUERY_DELIMITER = "&";
-  private static final String URL_QUERY_VALUE_DELIMITER = "=";
-
-  private static String getQueryDelimiter(String url) {
-    try {
-      return new URL(url).getQuery() != null ? URL_CONCAT_QUERY_DELIMITER
-          : URL_START_QUERY_DELIMITER;
-    } catch (MalformedURLException e) {
-      e.printStackTrace();
-    }
-    return null;
-  }
-
-  private static String addQueryParamsToUrl(List<Pair<String, String>> queryParams, String url) {
-    if (getQueryDelimiter(url) != null) {
-      url = url + getQueryDelimiter(url);
-
-      Iterator<Pair<String, String>> iterator = queryParams.iterator();
-      while (iterator.hasNext()) {
-        Pair<String, String> pair = iterator.next();
-        url =
-            url + pair.first + URL_QUERY_VALUE_DELIMITER + pair.second + URL_CONCAT_QUERY_DELIMITER;
-      }
-
-      return url.substring(0, url.length() - 2);
-    } else {
-      return null;
     }
   }
 
@@ -132,7 +97,7 @@ public class DeviceUtils {
 
             Ocm.getQueryStringGenerator().createQueryString(federatedAuthorization, queryString -> {
               if (queryString != null && !queryString.isEmpty()) {
-                String urlWithQueryParams = addQueryParamsToUrl(queryString, url);
+                String urlWithQueryParams = FAUtils.addQueryParamsToUrl(queryString, url);
                 System.out.println(
                     "Main ContentWebViewGridLayout federatedAuth url: " + urlWithQueryParams);
                 if (urlWithQueryParams != null) {
